Big Update ahead, transitioning to Cross platform
Switching working env to Linux and Rider, as part of the change, I'm making the system more
cross platform friendly. This means the old WPF project will be dead, but much of it
will be salvaged and repurposed for an Avalonia based UI.

Additionally, this transition allows the system to move from .Net Frawework to .Net Core.

Work will be needed to remake the Editor UI, but it should be relatively easy.